first stab at adding a pager format
This implements a third type of content which is :short_text. It is supposed
to be used for low character count messaging like SMS. It also cleans up some
things around initializing of subject lines before trying to print it and
checking that the mail object is not nil before trying to deliver on it.

Another big (but backwards compatible) change is the addition of methods to
delete sections from the @content hash. This means you can now call `super` in
your formatter methods and then just delete and reimplement the content you
actually care about. This makes it easier to implement specific formatters
without having to repeat code.